Synergy component folder renamed to conversation
