+++ app/src/components/conversation/avatar/Avatar.vue
