+++ app/src/components/conversation/percentage-ring/PercentageRing.vue
